// appfront/src/app/api/models/enrollment-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface EnrollmentResponse {
  completed?: boolean;
  courseId?: string;
  courseTitle?: string;
  enrollmentDate?: string;
  idEnrollment?: string;
  studentId?: string;
  studentName?: string;
  totalProgress?: number;
}